determines are now abstract
