+++ ts/block/basic_block.ts
@@ -63,7 +63,7 @@ export abstract class BasicBlock extends EventBlock {
 
     protected popup_event_eval() {
         Popup.remove_all_popup();
-        let block_eval_result = document.getElementById("block-eval-result"); 
+        let block_eval_result = document.getElementById("block-eval-result");
         if (block_eval_result !== null) {
             document.getElementById("roots").removeChild(block_eval_result);
         }
@@ -74,8 +74,8 @@ export abstract class BasicBlock extends EventBlock {
             case PyramidTypeID.Number:
                 result_block = new LiteralBlock(
                     [
-                        document.documentElement.clientWidth - this.playground.getBoundingClientRect().left -210,
-                        document.documentElement.clientHeight -210,
+                        document.documentElement.clientWidth - this.playground.getBoundingClientRect().left - 210,
+                        document.documentElement.clientHeight - 210,
                     ],
                     result.value,
                 );
